Move enums to a separate package
