--- calculator/src/main/java/org/smileksey/calculator/dto/ChangeType.java
@@ -1,4 +1,0 @@
-package org.smileksey.calculator.dto;
-
-public enum ChangeType {
-}
