+++ calculator/src/main/java/org/smileksey/calculator/dto/EmailMessage.java
@@ -2,6 +2,7 @@ package org.smileksey.calculator.dto;
 
 import lombok.Getter;
 import lombok.Setter;
+import org.smileksey.calculator.dto.enums.Theme;
 
 @Getter
 @Setter
