+++ calculator/src/main/java/org/smileksey/calculator/dto/EmploymentDto.java
@@ -2,6 +2,8 @@ package org.smileksey.calculator.dto;
 
 import lombok.Getter;
 import lombok.Setter;
+import org.smileksey.calculator.dto.enums.EmploymentStatus;
+import org.smileksey.calculator.dto.enums.Position;
 
 import java.math.BigDecimal;
