+++ calculator/src/main/java/org/smileksey/calculator/dto/ScoringDataDto.java
@@ -2,6 +2,8 @@ package org.smileksey.calculator.dto;
 
 import lombok.Getter;
 import lombok.Setter;
+import org.smileksey.calculator.dto.enums.Gender;
+import org.smileksey.calculator.dto.enums.MaritalStatus;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
