+++ calculator/src/main/java/org/smileksey/calculator/dto/StatementStatusHistoryDto.java
@@ -2,6 +2,8 @@ package org.smileksey.calculator.dto;
 
 import lombok.Getter;
 import lombok.Setter;
+import org.smileksey.calculator.dto.enums.ChangeType;
+import org.smileksey.calculator.dto.enums.Status;
 
 import java.time.LocalDateTime;
