--- calculator/src/main/java/org/smileksey/calculator/dto/Status.java
@@ -1,5 +1,0 @@
-package org.smileksey.calculator.dto;
-
-public enum Status {
-
-}
