--- calculator/src/main/java/org/smileksey/calculator/dto/Theme.java
@@ -1,4 +1,0 @@
-package org.smileksey.calculator.dto;
-
-public enum Theme {
-}
