+++ calculator/src/main/java/org/smileksey/calculator/dto/enums/EmploymentStatus.java
@@ -1,0 +1,4 @@
+package org.smileksey.calculator.dto.enums;
+
+public enum EmploymentStatus {
+}
