+++ calculator/src/main/java/org/smileksey/calculator/dto/enums/MaritalStatus.java
@@ -1,0 +1,4 @@
+package org.smileksey.calculator.dto.enums;
+
+public enum MaritalStatus {
+}
