+++ calculator/src/main/java/org/smileksey/calculator/dto/enums/Position.java
@@ -1,0 +1,4 @@
+package org.smileksey.calculator.dto.enums;
+
+public enum Position {
+}
