+++ calculator/src/main/java/org/smileksey/calculator/dto/enums/Status.java
@@ -1,0 +1,5 @@
+package org.smileksey.calculator.dto.enums;
+
+public enum Status {
+
+}
